fix some numeric types cast issues
Former-commit-id: 721aa3935dc8338494c60c00d7a4e48514f6f635
Former-commit-id: c436a313a5ad32f678664acb681c81be94d3bed3

+++ plugins/bullet/src/minko/component/bullet/ConeShape.hpp
@@ -85,7 +85,7 @@ namespace minko
 				volume() const
 				{
 					const float radius	= _radius + _margin; 
-					const float volume	= (PI / 3.0f) * radius * radius * (_height + _margin);
+					const float volume	= ((float)PI / 3.0f) * radius * radius * (_height + _margin);
 
 					return volume * _volumeScaling;
 				}
